requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.30.6
uagents==0.20.1
pydantic==2.8.2
httpx==0.26.0
anthropic==0.18.1
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.3
rich==13.7.0
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# ChromaDB and ML dependencies
chromadb==0.4.15
sentence-transformers==2.2.2
huggingface_hub==0.16.4
numpy==1.24.3

